refactor(gateway): isolate embeddings provider lifetime
Prepare the shared lifetime owner during startup without loading the embeddings HTTP request module. Preserve eager shutdown references and the existing admission, retirement, and drain behavior.
