+++ src/gateway/server-shutdown.runtime.test.ts
@@ -3,6 +3,7 @@ import { describe, expect, it, vi } from "vitest";
 const state = vi.hoisted(() => ({
   loaded: [] as string[],
   prepareClose: vi.fn(),
+  drainEmbeddingProviders: vi.fn(),
   completeClose: vi.fn(),
   flushSessionChanges: vi.fn(),
   stopPlugins: vi.fn(),
@@ -44,8 +45,11 @@ vi.mock("../agents/agent-bundle-lsp-runtime.js", () => {
   return { disposeAllBundleLspRuntimes: vi.fn() };
 });
 vi.mock("./embeddings-http.js", () => {
+  throw new Error("shutdown preparation must not load embeddings HTTP");
+});
+vi.mock("./embeddings-provider-lifetime.js", () => {
   state.loaded.push("embeddings");
-  return { drainRetainedOpenAiEmbeddingProviders: vi.fn() };
+  return { drainRetainedOpenAiEmbeddingProviders: state.drainEmbeddingProviders };
 });
 vi.mock("../hooks/gmail-watcher.js", () => {
   state.loaded.push("gmail-watcher");
@@ -90,6 +94,7 @@ describe("gateway shutdown runtime", () => {
       ].toSorted(),
     );
     expect(runtime.prepareGatewayClose).toBe(state.prepareClose);
+    expect(runtime.drainRetainedOpenAiEmbeddingProviders).toBe(state.drainEmbeddingProviders);
     expect(runtime.completeGatewayClose).toBe(state.completeClose);
     expect(runtime.flushPendingSessionsChangedEvents).toBe(state.flushSessionChanges);
     expect(runtime.runGlobalGatewayStopSafely).toBe(state.stopPlugins);
