+++ src/gateway/server-shutdown.runtime.ts
@@ -25,7 +25,7 @@ export async function prepareGatewayShutdownRuntime() {
     import("../tasks/task-registry.maintenance.js"),
     import("../agents/main-session-recovery/main-session-restart-recovery.js"),
     import("../agents/agent-bundle-lsp-runtime.js"),
-    import("./embeddings-http.js"),
+    import("./embeddings-provider-lifetime.js"),
     import("../hooks/gmail-watcher.js"),
     import("../agents/code-mode-state.js"),
     import("../agents/provider-transport-dispatcher-pool.js"),
